requirements.txt
requests
python-dotenv
wasabi
youtube_transcript_api